Moving logic from __main__ to __init__. We also need a "download()" function in __init__ that takes care of downloading.
