+++ semanticizest/parse_wikidump/__main__.py
@@ -19,7 +19,7 @@ from six.moves.urllib.request import urlretrieve
 import argparse
 from docopt import docopt
 
-from . import parse_dump
+from . import parse_dump, Db
 from .._semanticizer import createtables_path
 
 
@@ -37,41 +37,6 @@ class Progress(object):
         if done >= self.threshold * totalsize:
             logger.info("%3d%% done", int(100 * self.threshold))
             self.threshold += .05
-
-
-class Db(object):
-    def __init__(self, fname):
-        self.db_fname = fname
-        self.db = ""
-        
-    
-    def connect(self):
-        try:
-            self.db = sqlite3.connect(self.db_fname)
-        except sqlite3.OperationalError as e:
-            if 'unable to open' in str(e):
-                # This exception doesn't store the path.
-                die("%s: %r" % (e, self.db_fname))
-            else:
-                raise
-    
-    def disconnect(self):
-        if self.db:
-            self.db.close()
-    
-    def setup(self):
-        logger.info("Creating database at %r" % self.db_fname)
-        with open(createtables_path()) as f:
-            create = f.read()
-
-            c = self.db.cursor()
-            try:
-                c.executescript(create)
-            except sqlite3.OperationalError as e:
-                if re.search(r'table .* already exists', str(e)):
-                    die("database %r already populated" % self.db_fname)
-                else:
-                    raise
 
 
 DUMP_TEMPLATE = (
